+++ src/utils/index.js
@@ -2,3 +2,4 @@ import Console from './Console';
 import RandomNumber from './RandomNumber';
 
 export { Console, RandomNumber };
+export * from './validatorsUtils';
